Фикс обновления и сохранения выбранных офферов
