+++ src/utils/point.js
@@ -22,33 +22,6 @@ function calculatesTravelTime(dateFrom, dateTo) {
   return `${days ? `${days}D` : ''} ${hours ? `${hours}H` : ''} ${minutes}`;
 }
 
-
-function createFormOffersTemplate(pointOffers, point) {
-  return pointOffers
-    .map((offer) => {
-      const checked = point.includes(offer.id) ? 'checked' : '';
-      return `<div class="event__offer-selector">
-                        <input class="event__offer-checkbox  visually-hidden" id="${offer.id}" type="checkbox" name="event-offer-luggage" ${checked} ${point.isDisabled ? 'disabled' : ''}>
-                        <label class="event__offer-label" for="${offer.id}">
-                          <span class="event__offer-title">${offer.title}</span>
-                          &plus;&euro;&nbsp;
-                          <span class="event__offer-price">${offer.price}</span>
-                        </label>
-                      </div>`;
-    }).join('');
-}
-
-function createDestinationList(destinations) {
-  return destinations.map((destination) => `<option value="${destination.name}"></option>`).join('');
-}
-
-function createEventTypeItem (offers) {
-  return offers.map((offer) => `<div class="event__type-item">
-  <input id="event-type-${offer.type}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${offer.type}">
-  <label class="event__type-label  event__type-label--${offer.type}" for="event-type-${offer.type}-1">${(offer.type)[0].toUpperCase() + (offer.type).slice(1)}</label>
-</div>`).join('');
-}
-
 function isPointsPassed(dueDate) {
   const now = dayjs();
   return dayjs(dueDate).isBefore(now, 'day');
@@ -100,4 +73,4 @@ function isDatesEqual(dateA, dateB) {
   return (dateA === null && dateB === null) || dayjs(dateA).isSame(dateB, 'D');
 }
 
-export {humanizeTaskDueDate, calculatesTravelTime, createFormOffersTemplate, createDestinationList, createEventTypeItem, isPointsPassed, isPointsPlanned, isPointsCurrent, sortPointByDate, sortPointByPrice, sortPointByTime, isDatesEqual};
+export {humanizeTaskDueDate, calculatesTravelTime, isPointsPassed, isPointsPlanned, isPointsCurrent, sortPointByDate, sortPointByPrice, sortPointByTime, isDatesEqual};
