+++ src/view/edit-form-view.js
@@ -1,9 +1,41 @@
 import AbstractStatefulView from '../framework/view/abstract-stateful-view.js';
-import {createFormOffersTemplate, humanizeTaskDueDate, createDestinationList, createEventTypeItem} from '../utils/point.js';
+import {humanizeTaskDueDate} from '../utils/point.js';
 import {DATE_TIME_FORMAT} from '../const.js';
 import flatpickr from 'flatpickr';
 import 'flatpickr/dist/flatpickr.min.css';
 import he from 'he';
+
+function createDestinationList(destinations) {
+  return destinations.map((destination) => `<option value="${destination.name}"></option>`).join('');
+}
+
+function createFormOffersTemplate(pointOffers, point) {
+  return pointOffers
+    .map((offer) => {
+      const checked = point.includes(offer.id) ? 'checked' : '';
+      return `<div class="event__offer-selector">
+                        <input class="event__offer-checkbox  visually-hidden"
+                        id="event-offer-${offer.id}"
+                        data-offer-id="${offer.id}"
+                        type="checkbox"
+                        name="event-offer-${offer.type}-${offer.id}"
+                        ${checked}
+                        ${point.isDisabled ? 'disabled' : ''}>
+                        <label class="event__offer-label" for="event-offer-${offer.id}">
+                          <span class="event__offer-title">${offer.title}</span>
+                          &plus;&euro;&nbsp;
+                          <span class="event__offer-price">${offer.price}</span>
+                        </label>
+                      </div>`;
+    }).join('');
+}
+
+function createEventTypeItem (offers) {
+  return offers.map((offer) => `<div class="event__type-item">
+  <input id="event-type-${offer.type}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${offer.type}">
+  <label class="event__type-label  event__type-label--${offer.type}" for="event-type-${offer.type}-1">${(offer.type)[0].toUpperCase() + (offer.type).slice(1)}</label>
+</div>`).join('');
+}
 
 
 function createPictures(pictures) {
@@ -145,6 +177,10 @@ export default class EditFormView extends AbstractStatefulView {
     this.element.querySelector('.event__input--destination').addEventListener('change', this.#changeDestinationHandler);
     this.element.querySelector('.event__input--price').addEventListener('input', this.#changePriceHandler);
     this.element.querySelector('.event__reset-btn').addEventListener('click', this.#formDeleteClickHandler);
+    if (this.element.querySelector('.event__section--offers')) {
+      this.element.querySelector('.event__section--offers')
+        .addEventListener('change', this.#offerChangeHandler);
+    }
 
     this.#setDatepicker();
   }
@@ -173,6 +209,20 @@ export default class EditFormView extends AbstractStatefulView {
       destination: selectedDestination.id
     });
 
+  };
+
+  #offerChangeHandler = (evt) => {
+    const currentOffer = evt.target.dataset.offerId;
+
+    if (evt.target.checked) {
+      this._setState({
+        offers: [...this._state.offers, currentOffer]
+      });
+    } else {
+      this._setState({
+        offers: this._state.offers.filter((offer) => offer !== currentOffer)
+      });
+    }
   };
 
   #changePriceHandler = (evt) => {
